뷰 페이져 해결 (임시로 test a b c 창 띄우기)
해야될 것 : 리사이클러 뷰랑 연결
