+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.aze51.bidbid_client">
+    <uses-permission android:name="android.permission.INTERNET" />
+
 
     <application
         android:allowBackup="true"
@@ -8,7 +10,10 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity"></activity>
+        <activity android:name=".MainActivity">
+
+
+        </activity>
 
         <service android:name=".service.MyFirebaseMessagingService">
             <intent-filter>
@@ -31,5 +36,6 @@
         <activity android:name=".LoginActivity"></activity>
         <activity android:name=".JoinActivity"></activity>
     </application>
+
 
 </manifest>
