+++ app/src/main/java/com/aze51/bidbid_client/MainActivity.java
@@ -1,18 +1,24 @@
 package com.aze51.bidbid_client;
 
+import android.content.Context;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
 import android.support.v4.view.ViewPager;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
 
 import com.aze51.bidbid_client.Fragment.BottomMenuFragment;
-import com.aze51.bidbid_client.Fragment.ListFragment;
 import com.aze51.bidbid_client.Fragment.TitleFragment;
 import com.aze51.bidbid_client.Fragment.TopMenuFragment;
 import com.aze51.bidbid_client.ViewPager.ListItemData;
 import com.aze51.bidbid_client.ViewPager.RecyclerViewCustomAdapter;
+import com.aze51.bidbid_client.ViewPager.ViewPagerCustomAdapter;
 
 import java.util.ArrayList;
 
@@ -26,14 +32,19 @@ public class MainActivity extends AppCompatActivity {
     TopMenuFragment topMenuFragment;
     FragmentManager fragmentManager;
     TitleFragment titleFragment;
+    View rootViewBasic;
+
+
+    //ViewPager
+    ViewPager viewpager;
+
+
 
     //Recycler View
     ArrayList<ListItemData> itemDatas;
     RecyclerView recyclerView;
     RecyclerView.Adapter mAdapter;
     LinearLayoutManager mLayoutManager;
-
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -42,33 +53,29 @@ public class MainActivity extends AppCompatActivity {
         initiate();
         show_now_list();
     }
-
-
     //Made By Tae Joon 2016 06 27 : 현재 판매중인 목록 프래그먼트로 보여주기.
     private void show_now_list() {
-        recyclerView = (RecyclerView) findViewById(R.id.recyclerView_current);
+        fragmentManager.beginTransaction().add(R.id.TitleLayout, titleFragment).commit();
+        fragmentManager.beginTransaction().add(R.id.TopMenuLayout, topMenuFragment).commit();
+        fragmentManager.beginTransaction().add(R.id.ListLayout,listFragment).commit();
+        fragmentManager.beginTransaction().add(R.id.BottomLayout, bottomMenuFragment).commit();
+
+        /*recyclerView = (RecyclerView) findViewById(R.id.recyclerView_current);
+
         //아이템이 일정할 경우 고정
         recyclerView.setHasFixedSize(true);
         //LayoutManager 초기화
         mLayoutManager = new LinearLayoutManager(this);
         mLayoutManager.setOrientation(LinearLayoutManager.VERTICAL);
         recyclerView.setLayoutManager(mLayoutManager);
+
         //adapter 설정
         itemDatas = new ArrayList<ListItemData>();
         mAdapter = new RecyclerViewCustomAdapter(itemDatas);
         recyclerView.setAdapter(mAdapter);
 
-
-        itemDatas.add(new ListItemData(R.mipmap.b,"이름","가격","3:57 남음"));
-
-
-        fragmentManager.beginTransaction().add(R.id.TitleLayout, titleFragment).commit();
-        fragmentManager.beginTransaction().add(R.id.TopMenuLayout, topMenuFragment).commit();
-        fragmentManager.beginTransaction().add(R.id.ListLayout,listFragment).commit();
-        fragmentManager.beginTransaction().add(R.id.BottomLayout, bottomMenuFragment).commit();
+        itemDatas.add(new ListItemData(R.mipmap.b,"이름","가격","3:57 남음"));*/
     }
-
-
     //Made By Tae Joon 2016 06 27 : 초기화
     private void initiate() {
         bottomMenuFragment = new BottomMenuFragment();
@@ -76,5 +83,21 @@ public class MainActivity extends AppCompatActivity {
         topMenuFragment = new TopMenuFragment();
         fragmentManager = getSupportFragmentManager();
         titleFragment = new TitleFragment();
+    }
+    public class ListFragment extends Fragment { //view pager 사용해서 리사이클러 뷰 띄움
+
+        Context ctx;
+        public ListFragment(){
+            //생성자
+        }
+        @Nullable
+        @Override
+        public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+            rootViewBasic = inflater.inflate(R.layout.list_fragment,container,false);
+            viewpager = (ViewPager) rootViewBasic.findViewById(R.id.viewPager);
+            //ctx = getActivity().getApplicationContext();
+            viewpager.setAdapter(new ViewPagerCustomAdapter(reference));//Main Activity 의 this 를 보내야함.
+            return rootViewBasic;
+        }
     }
 }
