+++ app/src/main/java/com/aze51/bidbid_client/SplashActivity.java
@@ -23,6 +23,6 @@ public class SplashActivity extends AppCompatActivity {
                 startActivity(intent);
                 finish();
             }
-        }, 1500);
+        }, 700);
     }
 }
