+++ app/src/main/res/layout/current_list.xml
@@ -9,10 +9,5 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="current"/>
-
     </android.support.v7.widget.RecyclerView>
 </RelativeLayout>
