+++ app/src/main/res/layout/scheduled_list.xml
@@ -6,6 +6,5 @@
         android:id="@+id/recyclerView_scheduled"
         android:layout_width="match_parent"
         android:layout_height="match_parent">
-
     </android.support.v7.widget.RecyclerView>
 </RelativeLayout>
